FileSystem library improved

+++ src/FileSystem.php
@@ -4,7 +4,9 @@
 namespace Pechynho\Utility;
 
 
+use DirectoryIterator;
 use InvalidArgumentException;
+use RuntimeException;
 
 /**
  * @author Jan Pech <pechynho@gmail.com>
@@ -50,16 +52,23 @@ class FileSystem
 		if (FileSystem::isFile($source))
 		{
 			if (FileSystem::isFile($destination) && !$overwrite) throw new InvalidArgumentException("File '$destination' already exists.");
-			copy($source, $destination);
+			if (copy($source, $destination) === false)
+			{
+				throw new RuntimeException(sprintf('Could not copy file from %s to %s.', $source, $destination));
+			}
 			return;
 		}
 		if (FileSystem::isDirectory($destination) && !$overwrite) throw new InvalidArgumentException("Directory '$destination' already exists.");
 		if (!FileSystem::isDirectory($destination)) FileSystem::createDirectory($destination);
-		$items = array_diff(scandir($source), [".", ".."]);
-		foreach ($items as $item)
+		$iterator = new DirectoryIterator($source);
+		foreach ($iterator as $item)
 		{
-			$sourceItem = FileSystem::combinePath($source, $item);
-			$destinationItem = FileSystem::combinePath($destination, $item);
+			if ($item->isDot())
+			{
+				continue;
+			}
+			$sourceItem = FileSystem::combinePath($source, $item->getFilename());
+			$destinationItem = FileSystem::combinePath($destination, $item->getFilename());
 			FileSystem::copy($sourceItem, $destinationItem, $overwrite);
 		}
 	}
@@ -106,7 +115,10 @@ class FileSystem
 		}
 		if (!FileSystem::isDirectory($directory))
 		{
-			mkdir($directory, $mode, true);
+			if (mkdir($directory, $mode, true) === false)
+			{
+				throw new RuntimeException(sprintf("Could not create directory %s.", $directory));
+			}
 		}
 	}
 
@@ -126,6 +138,10 @@ class FileSystem
 			$finalPath = $finalPath . ($index == 0 ? "" : "/") . $path;
 		}
 		$finalPath = preg_replace('/[\/]{2,}/', '/', $finalPath);
+		if (false !== $realpath = realpath($finalPath))
+		{
+			$finalPath = $realpath;
+		}
 		return $finalPath;
 	}
 
@@ -160,16 +176,20 @@ class FileSystem
 		{
 			throw new InvalidArgumentException("File or directory '$destination' already exists.");
 		}
-		rename($source, $destination);
+		if (rename($source, $destination) === false)
+		{
+			throw new RuntimeException(sprintf("Could not rename %s to %s", $source, $destination));
+		}
 	}
 
 	/**
 	 * @param string $filename
-	 * @param mixed  $content
+	 * @param string $content
 	 * @param bool   $overwrite
 	 */
 	public static function write($filename, $content, $overwrite = false)
 	{
+		ParamsChecker::isString('$content', $content, __METHOD__);
 		if (!is_string($filename))
 		{
 			throw new InvalidArgumentException('Parameter $filename has to be type of string.');
@@ -187,10 +207,19 @@ class FileSystem
 			throw new InvalidArgumentException("File '$filename' already exists.");
 		}
 		if (FileSystem::isFile($filename)) FileSystem::delete($filename);
-		$file = fopen($filename, "w");
-		fwrite($file, $content);
-		fclose($file);
-		clearstatcache();
+		if (false === $file = fopen($filename, "w"))
+		{
+			throw new RuntimeException(sprintf("Function fopen('%s', 'w') failed.", $filename));
+		}
+		if (fwrite($file, $content) === false)
+		{
+			throw new RuntimeException(sprintf("Could not write (fwrite) content to %s.", $filename));
+		}
+		if (fclose($file) === false)
+		{
+			throw new RuntimeException(sprintf('Could not close (fclose) file %s.', $filename));
+		}
+		clearstatcache(true, $filename);
 	}
 
 	/**
@@ -208,25 +237,36 @@ class FileSystem
 		}
 		if (FileSystem::isFile($path))
 		{
-			unlink($path);
+			if (unlink($path) === false)
+			{
+				throw new RuntimeException(sprintf("Could not delete file %s.", $path));
+			}
 			return;
 		}
-		$items = array_diff(scandir($path), [".", ".."]);
-		foreach ($items as $item)
+		$iterator = new DirectoryIterator($path);
+		foreach ($iterator as $item)
 		{
-			$item = FileSystem::combinePath($path, $item);
+			if ($item->isDot())
+			{
+				continue;
+			}
+			$item = FileSystem::combinePath($path, $item->getFilename());
 			FileSystem::delete($item);
 		}
-		rmdir($path);
+		if (rmdir($path) === false)
+		{
+			throw new RuntimeException(sprintf("Could not delete file %s.", $path));
+		}
 	}
 
 	/**
 	 * @param string $filename
-	 * @param mixed  $content
+	 * @param string $content
 	 * @param bool   $newLine
 	 */
 	public static function append($filename, $content, $newLine = true)
 	{
+		ParamsChecker::isString('$content', $content, __METHOD__);
 		if (!is_string($filename))
 		{
 			throw new InvalidArgumentException('Parameter $filename has to be type of string.');
@@ -239,21 +279,36 @@ class FileSystem
 		{
 			throw new InvalidArgumentException("Given value '$filename' is not valid filename.");
 		}
-		$file = fopen($filename, "a");
+		if (false === $file = fopen($filename, "a"))
+		{
+			throw new RuntimeException(sprintf("Could not open file fopen('%s', 'a').", $filename));
+		}
 		if (FileSystem::isEmpty($filename))
 		{
-			fwrite($file, $content);
+			if (fwrite($file, $content) === false)
+			{
+				throw new RuntimeException(sprintf("Could not write content to file %s.", $filename));
+			}
 		}
 		else if (!$newLine)
 		{
-			fwrite($file, $content);
+			if (fwrite($file, $content) === false)
+			{
+				throw new RuntimeException(sprintf("Could not write content to file %s.", $filename));
+			}
 		}
 		else
 		{
-			fwrite($file, PHP_EOL . $content);
+			if (fwrite($file, PHP_EOL . $content) === false)
+			{
+				throw new RuntimeException(sprintf("Could not write content to file %s.", $filename));
+			}
 		}
-		fclose($file);
-		clearstatcache();
+		if (fclose($file) === false)
+		{
+			throw new RuntimeException(sprintf('Could not close (fclose) file %s.', $filename));
+		}
+		clearstatcache(true, $filename);
 	}
 
 	/**
@@ -287,12 +342,23 @@ class FileSystem
 		{
 			throw new InvalidArgumentException("Path '$path' does not exist.");
 		}
-		if (FileSystem::isFile($path)) return filesize($path);
-		$items = array_diff(scandir($path), [".", ".."]);
-		$size = 0;
-		foreach ($items as $item)
+		if (FileSystem::isFile($path))
 		{
-			$item = FileSystem::combinePath($path, $item);
+			if (false === $size = filesize($path))
+			{
+				throw new RuntimeException(sprintf("Could not read file size of %s.", $path));
+			}
+			return $size;
+		}
+		$iterator = new DirectoryIterator($path);
+		$size = 0;
+		foreach ($iterator as $item)
+		{
+			if ($item->isDot())
+			{
+				continue;
+			}
+			$item = FileSystem::combinePath($path, $item->getFilename());
 			$size = $size + FileSystem::size($item);
 		}
 		return $size;
@@ -317,14 +383,24 @@ class FileSystem
 		{
 			throw new InvalidArgumentException("File '$filename' does not exist.");
 		}
-		$file = fopen($filename, "r");
+		if (false === $file = fopen($filename, "r"))
+		{
+			throw new RuntimeException(sprintf("Could not open file fopen('%s', 'r').", $filename));
+		}
 		$lines = [];
 		while (!feof($file))
 		{
-			$lines[] = $trimEndOfLine ? Strings::trimEnd(fgets($file), [PHP_EOL]) : fgets($file);
+			if (false === $line = fgets($file))
+			{
+				throw new RuntimeException(sprintf("Could not read line (fgets) of %s", $filename));
+			}
+			$lines[] = $trimEndOfLine ? Strings::trimEnd($line, [PHP_EOL]) : $line;
 		}
-		fclose($file);
-		clearstatcache();
+		if (fclose($file) === false)
+		{
+			throw new RuntimeException(sprintf('Could not close (fclose) file %s.', $filename));
+		}
+		clearstatcache(true, $filename);
 		return $lines;
 	}
 
@@ -369,13 +445,23 @@ class FileSystem
 		{
 			throw new InvalidArgumentException("File '$filename' does not exist.");
 		}
-		$file = fopen($filename, "r");
+		if (false === $file = fopen($filename, "r"))
+		{
+			throw new RuntimeException(sprintf("Could not open file fopen('%s', 'r').", $filename));
+		}
 		while (!feof($file))
 		{
-			yield $trimEndOfLine ? Strings::trimEnd(fgets($file), [PHP_EOL]) : fgets($file);
+			if (false === $line = fgets($file))
+			{
+				throw new RuntimeException(sprintf("Could not read line (fgets) of %s", $filename));
+			}
+			yield $trimEndOfLine ? Strings::trimEnd($line, [PHP_EOL]) : $line;
 		}
-		fclose($file);
-		clearstatcache();
+		if (fclose($file) === false)
+		{
+			throw new RuntimeException(sprintf('Could not close (fclose) file %s.', $filename));
+		}
+		clearstatcache(true, $filename);
 	}
 
 	/**
@@ -399,10 +485,14 @@ class FileSystem
 			throw new InvalidArgumentException('Parameter $recursively has to be type of boolean.');
 		}
 		$output = [];
-		$items = array_diff(scandir($directory), [".", ".."]);
-		foreach ($items as $item)
+		$iterator = new DirectoryIterator($directory);
+		foreach ($iterator as $item)
 		{
-			$item = FileSystem::combinePath($directory, $item);
+			if ($item->isDot())
+			{
+				continue;
+			}
+			$item = FileSystem::combinePath($directory, $item->getFilename());
 			if (FileSystem::isFile($item) && ($mode == FileSystem::SCAN_FILES || $mode == FileSystem::SCAN_ALL))
 			{
 				$output[] = $item;
